fix clippy lints

+++ src/tests/good_tests/ion_parser_good_serde_value.rs
@@ -34,7 +34,7 @@ fn serde_from_ion_big_integer() {
     let big_int = BigInt::from_str("-201545002204320").unwrap();
     let big_int_ion = IonValue::BigInteger(big_int);
     let serde_big_int: Value = big_int_ion.try_into().unwrap();
-    let number = -201545002204320 as i64;
+    let number = -201545002204320_i64;
 
     assert_eq!(serde_big_int, Value::Number(number.into()));
 }
